+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - More strict type for useContextUpdate return value
+- Fallback missing runWithPriority for preact (#36)
 
 ## [1.3.4] - 2020-12-12
 ### Changed